A behaviour has been added to allow add additional actions after an exception occurs
